refactor: :recycle: asset info reorganizzation
