--- packages/web/components/token-details/index.ts
@@ -1,1 +1,0 @@
-export * from "./token-details";
