--- packages/web/components/twitter-section/index.ts
@@ -1,1 +1,0 @@
-export * from "./twitter-section";
