+++ packages/web/hooks/use-asset-info-view.ts
@@ -1,0 +1,11 @@
+import { ObservableAssetInfoConfig } from "~/hooks/ui-config";
+import { createContext } from "~/utils/react-context";
+
+const [AssetInfoViewProvider, useAssetInfoView] = createContext<{
+  assetInfoConfig: ObservableAssetInfoConfig;
+}>({
+  name: "AssetInfoViewContext",
+  strict: true,
+});
+
+export { AssetInfoViewProvider, useAssetInfoView };
